🔥 chore: move directory. pages to src/pages
